% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRT.R
\name{calcHRTParams}
\alias{calcHRTParams}
\alias{calcHRTParams,HRT-method}
\title{Calculate HRT parameters}
\usage{
calcHRTParams(HRTObj, IL = c_normIL, normIL = c_normIL)

\S4method{calcHRTParams}{HRT}(HRTObj, IL = c_normIL, normIL = c_normIL)
}
\arguments{
\item{HRTObj}{(HRT) The HRT object of which the parameters should be calculated}

\item{IL}{(Numeric) The overall arithmetic mean of the interval length of the
measurement to normalise TS}

\item{normIL}{(Numeric) The interval length to which TS should be normalised}
}
\value{
(HRT) An HRT object with (re)calculated HRT parameters
}
\description{
Calculates all HRT parameters needed for an HRT object
and saves them in the corresponding slots.
}
\details{
This method is a wrapper for the methods calcTO and calcTS.
}
