% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRT.R
\name{calcTO}
\alias{calcTO}
\alias{calcTO,HRT-method}
\title{Calculate TO parameters}
\usage{
calcTO(HRTObj)

\S4method{calcTO}{HRT}(HRTObj)
}
\arguments{
\item{HRTObj}{(HRT) The HRT object, for which TO should be calculated}
}
\value{
(HRT) An HRT object with (re)calculated TO
}
\description{
Calculates the TO parameters and saves it in the corresponding slot
}
