% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{checkAnnotations}
\alias{checkAnnotations}
\title{Checks annotations for compatibility}
\usage{
checkAnnotations(annotations, input, PVCAnn, label)
}
\arguments{
\item{annotations}{Alphabetical vector}

\item{input}{Numeric vector}

\item{PVCAnn}{Character}

\item{label}{Name of the data given and formatted for output}
}
\description{
Checks annotations for compatibility
}
