% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{checkForHRT}
\alias{checkForHRT}
\title{Checks RR-intervals for HRT criteria and returns an HRT object}
\usage{
checkForHRT(intervals, numPreRRs = c_numPreRRs, numPostRRs = c_numPostRRs)
}
\arguments{
\item{intervals}{Numeric vector}

\item{numPreRRs}{Numeric}

\item{numPostRRs}{Numeric}
}
\value{
HRT A single HRT object
}
\description{
Checks RR-intervals for HRT criteria and returns an HRT object
}
