% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{checkInput}
\alias{checkInput}
\title{Checks data input for compatibility}
\usage{
checkInput(input, numSnippet, label)
}
\arguments{
\item{input}{Numeric vector}

\item{numSnippet}{Numeric, number of RRs in the the HRT snippet}

\item{label}{Name of the data given and formatted for output}
}
\description{
Checks data input for compatibility
}
