% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{checkIntervals}
\alias{checkIntervals}
\title{Checks interval data for validity}
\usage{
checkIntervals(input, label)
}
\arguments{
\item{input}{(Numeric vector) RR intervals in ms}

\item{label}{(Character) Name of the data given and formatted for output}
}
\description{
Checks interval data for validity
}
