% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRTList.R
\name{checkValidity}
\alias{checkValidity}
\alias{checkValidity,HRTList-method}
\title{Check for HRTList class}
\usage{
checkValidity(x, ...)

\S4method{checkValidity}{HRTList}(x, av = FALSE, pos = FALSE)
}
\arguments{
\item{x}{HRTList}

\item{av}{Boolean Should avHRT be checked?}

\item{pos}{Boolean Should pos be checked?}
}
\description{
Check for HRTList class
}
