% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{cleanInput}
\alias{cleanInput}
\title{Cleans data input for further checks or calculation}
\usage{
cleanInput(input)
}
\arguments{
\item{input}{(Numeric vector) RR intervals in ms}
}
\value{
(numeric vector) Input vector without possible bias
}
\description{
Cleans data input for further checks or calculation
}
