% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRTList.R
\name{getHRTParams}
\alias{getHRTParams}
\alias{getHRTParams,HRTList-method}
\title{Extracts all values of a special slot out of a HRTList}
\usage{
getHRTParams(HRTListObj, sl)

\S4method{getHRTParams}{HRTList}(HRTListObj, sl)
}
\arguments{
\item{HRTListObj}{HRTList object}

\item{sl}{Character, choose a slot saved by an HRT object}
}
\value{
Vector
}
\description{
Extracts all values of the given slot in each HRT of the HRTList
and returns them in a list
}
