% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{getHRTs}
\alias{getHRTs}
\title{Finds HRTs}
\usage{
getHRTs(
  intervals,
  annotations = NULL,
  PVCAnn = "V",
  numPreRRs = c_numPreRRs,
  numPostRRs = c_numPostRRs,
  numSnippet
)
}
\arguments{
\item{intervals}{Numeric vector}

\item{annotations}{Alphabetical vector}

\item{PVCAnn}{Character}

\item{numPreRRs}{Numeric}

\item{numPostRRs}{Numeric}

\item{numSnippet}{Numeric, number of RRs in the HRT snippet}
}
\value{
HRTListObj
}
\description{
Scans for HRTs in the given vector and returns an HRTList object.
}
