% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRTList.R
\name{getPositions}
\alias{getPositions}
\alias{getPositions,HRTList-method}
\title{Get positions of PVCs}
\usage{
getPositions(HRTListObj)

\S4method{getPositions}{HRTList}(HRTListObj)
}
\arguments{
\item{HRTListObj}{(HRTList object)}
}
\value{
No return value, possibly throws errors/warnings
}
\description{
Returns the positions of all ventricular premature complexes (VPCs) and
accordingly the coupling intervals that were found in the given vector
when the HRTList was created.
}
