% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRT.R
\name{getRRs}
\alias{getRRs}
\alias{getRRs,HRT-method}
\title{Returns the VPCS intervals in right order}
\usage{
getRRs(HRTObj)

\S4method{getRRs}{HRT}(HRTObj)
}
\arguments{
\item{HRTObj}{HRT}
}
\description{
Returns the VPCS intervals in right order
}
