% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRTList.R
\docType{class}
\name{HRTList}
\alias{HRTList}
\title{S4 class to represent a list of HRT objects}
\description{
This class specifies an object to save all HRT objects of a given vector. It
also saves an averaged HRT for calculation of the averaged HRT parameters and
plotting of all HRTs in a single plot.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character, Name of the vector if given}

\item{\code{IL}}{Numeric, Arithmetic mean of the overall interval length of the vector}

\item{\code{pos}}{Numeric vector, Positions of premature ventricular complexes in
given input}

\item{\code{HRTs}}{List, all HRT objects}

\item{\code{avHRT}}{avHRT object, the average of all HRTs}

\item{\code{RMSSD}}{Numeric, square root of the mean of the squared successive
differences between adjacent intervals of the whole measurement}
}}

\note{
After using \code{vectorToHRT} all slots in the resulting HRTList
object are set. Please do not set them manually since many functions of the
HRTList class rely on valid values assigned to the needed slots.
}
