% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRT.R
\name{plot,HRT-method}
\alias{plot,HRT-method}
\title{Plot an HRT object}
\usage{
\S4method{plot}{HRT}(
  x,
  cropped = TRUE,
  add = FALSE,
  TT = FALSE,
  pch = 20,
  xlab = "# of RR interval",
  ylab = "length of RR interval (ms)",
  paramsLegend = TRUE,
  colTO = "#ec2023",
  colTS = "#006AFF",
  colTT = "#6800DE",
  ...
)
}
\arguments{
\item{x}{HRT, A HRT object}

\item{cropped}{Boolean, Should the plot be cut to focuse on the HRT parameters?
To show all points use FALSE.}

\item{add}{Boolean, Should the given HRT be added to a plot?}

\item{TT}{Boolean, Should Turbulence timing be marked?}

\item{pch}{Numeric, Plotting character, for other options see graphics::var}

\item{xlab}{Character, Label for the x axis}

\item{ylab}{Character Label for the x axis}

\item{paramsLegend}{Boolean, should the parameter values of the HRT be plotted?}

\item{colTO}{Character, colour used to highlight TO}

\item{colTS}{Character, colour used to highlight TS}

\item{colTT}{Character, colour used to highlight TT}

\item{...}{Other arguments in tag = value form. See graphics::par for more information.}
}
\description{
Plots RR-intervals saved in the HRT object and marks
turbulence onset and turbulence slope.
}
\note{
Please note that the argument xaxt and ylim can't be set,
 since the axis as well as the ranges of the y axis are set by the function.
}
