% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRTList.R
\name{plot,HRTList-method}
\alias{plot,HRTList-method}
\title{Plot an HRT object}
\usage{
\S4method{plot}{HRTList}(
  x,
  cropped = TRUE,
  TT = FALSE,
  pch = 20,
  xlab = "# of RR interval",
  ylab = "length of RR interval (ms)",
  paramsLegend = TRUE,
  colTO = "#ec2023",
  colTS = "#006AFF",
  colTT = "#6800DE",
  ...
)
}
\arguments{
\item{x}{HRTList}

\item{cropped}{(Boolean) Should the plot be cut to focus on the HRT parameters?
To show all points use FALSE.}

\item{TT}{(Boolean) Should Turbulence timing be marked?}

\item{pch}{(Numeric) Plotting character, for other options see graphics::var}

\item{xlab}{(Character) Label for the x axis}

\item{ylab}{(Character) Label for the x axis}

\item{paramsLegend}{(Boolean) Should the parameter values of the HRT be plotted?}

\item{colTO}{(Character) Colour used to highlight TO}

\item{colTS}{(Character) Colour used to highlight TS}

\item{colTT}{(Character) Colour used to highlight TT}

\item{...}{Other arguments in tag = value form}
}
\value{
No return value
}
\description{
Plots RR-intervals saved in the HRT object and marks HRT parameters.
}
\note{
Please note that the argument xaxt and ylim can't be set,
 since the axis as well as the ranges of the y axis are set by the function.
}
