% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataLong}
\alias{testdataLong}
\title{Long term data}
\format{
A numeric vector.
}
\usage{
testdataLong
}
\description{
Interval data provided by Physionet (001 of NSR2DB). This dataset
contains a full long term measurement with several HRTs.
}
\keyword{datasets}
