% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataRegular}
\alias{testdataRegular}
\title{Intervals without noise}
\format{
A numeric vector of length 1000.
}
\usage{
testdataRegular
}
\description{
Artificial dummy interval data: This dataset includes 5 identical VPCSs 
that fit the HRT filter rules and was created without noise.
}
\keyword{datasets}
