% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataRegular_HRTObj}
\alias{testdataRegular_HRTObj}
\title{Intervals without noise}
\format{
An HRTList object.
}
\usage{
testdataRegular_HRTObj
}
\description{
HRTList Object that was generated from testdataRegular and includes 5
identical HRTs. For more information see the vignette.
}
\keyword{datasets}
