% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataVariant}
\alias{testdataVariant}
\title{Intervals with noise}
\format{
A numeric vector of length 1000.
}
\usage{
testdataVariant
}
\description{
Artificial dummy interval data: This dataset includes 5 different VPCSs 
that fit the HRT filter rules.
}
\keyword{datasets}
