% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataVariant_HRTObj}
\alias{testdataVariant_HRTObj}
\title{Intervals with noise}
\format{
An HRTList object.
}
\usage{
testdataVariant_HRTObj
}
\description{
HRTList Object that was generated from testdataVariant and includes 3 HRTs
based on data from Physionet (NSRDB). For more information see the vignette.
}
\keyword{datasets}
