% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataVariantNoHRT_HRTObj}
\alias{testdataVariantNoHRT_HRTObj}
\title{Intervals with noise and without HRT}
\format{
An empty HRTList object.
}
\usage{
testdataVariantNoHRT_HRTObj
}
\description{
HRTList Object that was generated from testdataVariantNoHRT. This HRTList
does not include any HRT.
}
\keyword{datasets}
