% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{vectorToHRT}
\alias{vectorToHRT}
\title{Convert a vector to HRTList}
\usage{
vectorToHRT(
  input,
  annotations = NULL,
  PVCAnn = "V",
  normIL = c_normIL,
  normHallstrom = TRUE,
  numPreRRs = c_numPreRRs,
  numPostRRs = c_numPostRRs,
  inputName = as.character(NA),
  minHRT = 5,
  cleaning = TRUE
)
}
\arguments{
\item{input}{Numeric vector}

\item{annotations}{Alphabetical vector}

\item{PVCAnn}{Character}

\item{normIL}{Numeric, The interval length to which TS should be normalised}

\item{normHallstrom}{Boolean, should the normalisation of Hallstrom be used?}

\item{numPreRRs}{Numeric}

\item{numPostRRs}{Numeric}

\item{inputName}{String, name of the data}

\item{minHRT}{Numeric, minimal number of HRTs that are needed to create a HRTList object}

\item{cleaning}{Boolean, Should the input be roughly cleaned from artefacts before calculating IL and RMSSD?}
}
\value{
HRTList HRTList object
}
\description{
Scans for heart rate turbulence in a vector of RR-intervals and returns an
HRTList object including all found HRT objects.
The HRT criteria used were published by Schmidt et al.
(more information can be found in the vignette.)
}
