% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapply.R
\name{wapply}
\alias{wapply}
\title{Apply method on sliding window}
\source{
https://rmazing.wordpress.com/2013/04/23/wapply-a-faster-but-less-functional-rollapply-for-vector-setups/
}
\usage{
wapply(x, width, by = NULL, FUN = NULL, ...)
}
\arguments{
\item{x}{Vector}

\item{width}{windowsize}

\item{by}{Value how much to slide}

\item{FUN}{function to be applied}

\item{...}{additional arguments for FUN}
}
\value{
vector with return values of FUN for each window
}
\description{
Applies a given function on a vector by rolling over it with a sliding window 
mechanism.
}
\author{
A. N. Spiess, University Hospital Hamburg-Eppendorf
}
