% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{checkAnnotations}
\alias{checkAnnotations}
\title{Checks annotations for compatibility}
\usage{
checkAnnotations(annotations, input, PVCAnn, label)
}
\arguments{
\item{annotations}{(Character vector) Annotations matching input}

\item{input}{(Numeric vector) RR intervals in ms}

\item{PVCAnn}{(Character) Character that marks a VPC in the annotations}

\item{label}{(Character) Name of the data given and formatted for output}
}
\value{
No return value, possibly throws errors/warnings
}
\description{
Checks annotations for compatibility
}
