% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{checkInput}
\alias{checkInput}
\title{Checks data input for compatibility}
\usage{
checkInput(input, numSnippet, label)
}
\arguments{
\item{input}{(Numeric vector) RR intervals in ms}

\item{numSnippet}{(Numeric) number of RRs in the the HRT snippet}

\item{label}{(Character) Name of the data given and formatted for output}
}
\value{
No return value, possibly throws errors/warnings
}
\description{
Checks data input for compatibility
}
