% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_to_HRT.R
\name{getHRTs}
\alias{getHRTs}
\title{Finds HRTs}
\usage{
getHRTs(
  intervals,
  annotations = NULL,
  PVCAnn = "V",
  numPreRRs = c_numPreRRs,
  numPostRRs = c_numPostRRs,
  numSnippet
)
}
\arguments{
\item{intervals}{(Numeric vector) RR intervals in ms}

\item{annotations}{(Character vector) Annotations matching input}

\item{PVCAnn}{(Character) Character that marks a VPC in the annotations}

\item{numPreRRs}{(Numeric) Number of RRs before the coupling interval that are used for filtering}

\item{numPostRRs}{(Numeric) Number of RRs after the compensatory interval that are used for filtering}

\item{numSnippet}{(Numeric) Number of RRs in the HRT snippet}
}
\value{
(HRTList) HRTList with only pos and HRTs set
}
\description{
Scans for HRTs in the given vector and returns an HRTList object.
}
