% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRTList.R
\name{plot,HRTList-method}
\alias{plot,HRTList-method}
\title{Plot an HRT object}
\usage{
\S4method{plot}{HRTList}(x, y, ...)
}
\arguments{
\item{x}{HRTList}

\item{...}{Other arguments in tag = value form}
}
\description{
Plots RR-intervals saved in the HRT object and marks HRT parameters.
}
