% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataLong}
\alias{testdataLong}
\title{Long term data}
\format{
A numeric vector.
}
\usage{
testdataLong
}
\description{
Artificial dummy interval data: This dataset represents a long-term
measurement and includes 15 VPCSs that fit the HRT filter rules.
}
\keyword{datasets}
