% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataLong_Ann}
\alias{testdataLong_Ann}
\title{Long term data annotations}
\format{
A vector of characters.
}
\usage{
testdataLong_Ann
}
\description{
Artificial dummy interval data: This dataset contains the annotations
matching testdataLong.
}
\keyword{datasets}
