% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataVariant}
\alias{testdataVariant}
\title{Intervals with noise}
\format{
A numeric vector of length 1000.
}
\usage{
testdataVariant
}
\description{
Interval data based on a dataset provided by Physionet (NSRDB). This dataset
includes 3 HRTs that fit the HRT filter rules. For more information see the
vignette
}
\keyword{datasets}
