% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testdata-description.R
\docType{data}
\name{testdataVariantNoHRT}
\alias{testdataVariantNoHRT}
\title{Intervals with noise and without HRT}
\format{
A numeric vector of length 1000.
}
\usage{
testdataVariantNoHRT
}
\description{
Artificial dummy interval data: This dataset does not include any VPCSs.
}
\keyword{datasets}
